Revert "Changes in the UI of Chat Navbar"
This reverts commit 63a4901996b56501881d13db0547b2e0df2ff029.
