--- client/app/components/chatsNavigation/chatsNav.module.scss
@@ -1,100 +1,0 @@
-// Main Container
-.messageNavContainer {
-    display: flex;
-    flex-direction: column;
-    width: 25%;
-}
-
-// Header for the chats nav
-.messageHeader {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    h3 {
-        font-size: 2em;
-        font-weight: 600;
-        letter-spacing: 1px;
-    }
-
-    button {
-        // all: unset;
-        border: none;
-        // padding: 8px
-        height: 40px;
-        font-weight: 100;
-        width: 40px;
-        border-radius: 30%;
-        font-size: 1.7em;
-        background-color: white;
-        color: var(--dark-grey);
-        cursor: pointer;
-        box-shadow: 1px 1px 2px var(--dark-grey);
-    }
-}
-
-// Search Chats Box
-.searchBox {
-    width: 100%;
-    height: 5vh;
-    display: flex;
-    justify-content: center;
-    margin-bottom: 10px;
-    
-    input {
-        width: 100%;
-        border: none;
-        border-radius: 8px;
-        box-shadow: 1px 1px 2px var(--dark-grey);
-        padding: 0 15px;
-        font-size: 1em;
-
-        &:focus-visible {
-            outline: none;
-        }
-    }
-}
-
-// Chats in the nav
-.chatWrapper {
-    display: flex;
-    flex-direction: column;
-    margin-top: 30px;
-    .chatPersonContainer {
-        padding: 15px 10px;
-        margin-bottom: 2px;
-        cursor: pointer;
-    }
-    .chatPerson {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-    }
-    .selectedChat {
-        background-color: white;
-        border-radius: 5px;
-        box-shadow: 4px 4px 10px var(--dark-grey);
-    }
-    .nameAndAvatar {
-        display: flex;
-        gap: 10px;
-    }
-    .nameAndMessage {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        gap: 2.5px;
-    }
-    .chatName {
-        font-weight: 500;
-    }
-    .newMessage {
-        font-size: 0.9em;
-    }
-    .avatar {
-        border-radius: 50%;
-    }
-    .chatTime {
-        font-size: 0.9em;
-    }
-}
