+++ client/app/page.tsx
@@ -1,16 +1,13 @@
-import ChatsNav from "@/app/components/chatsNavigation/ChatsNav";
+import ChatsNav from "@/components/chatsNavigation/ChatsNav";
 import styles from "./page.module.scss";
-import ChatSpace from "@/app/components/chatSpace/ChatSpace";
-import { RoomContextProvider } from "./context/RoomContext";
+import ChatSpace from "@/components/chatSpace/ChatSpace";
 
 export default function Home() {
   return (
     <main className={styles.main}>
-      <RoomContextProvider>
-        <ChatsNav />
-        <hr />
-        <ChatSpace />
-      </RoomContextProvider>
+      <ChatsNav />
+      <hr />
+      <ChatSpace />
     </main>
   );
 }
