+++ client/components/chatSpace/ChatSpace.tsx
@@ -4,7 +4,6 @@ import React, { createRef, useRef } from "react";
 import styles from "./chatSpace.module.scss";
 import { useEffect, useState } from "react";
 import socket from "@/socket";
-import { useRoomContext } from "@/app/context/RoomContext";
 
 type messageObj = {
   text: string;
@@ -14,17 +13,16 @@ type messageObj = {
 const ChatSpace = () => {
   const [input, setInput] = useState("");
   const [messages, setMessages] = useState<Array<messageObj>>([]);
-  const { state } = useRoomContext();
 
   useEffect(() => {
     socket.on("connect", () => {
-      console.log("connected", socket.id);
+      console.log("connected");
     });
 
     return () => {
-      socket.on('disconnect', () => {
-        console.log("disconnected", socket.id)
-      })
+      socket.on("disconnect", () => {
+        console.log("disconnected");
+      });
     };
   }, []);
   
@@ -37,9 +35,8 @@ const ChatSpace = () => {
   }
 
   function handleClick() {
-    console.log(state.currentRoom)
     if (input.length > 0) {
-      socket.emit("send_message", {input, id: state.currentRoom});
+      socket.emit("send_message", input);
       setMessages([...messages, { text: input, type: "sent" }]);
       setInput("");
     } else {
