+++ client/components/chatSpace/chatSpace.module.scss
@@ -34,8 +34,8 @@
 
     div {
         width: fit-content;
-        font-size: 1em;
-        padding: 10px 18px;
+        font-size: 1.1em;
+        padding: 7px 15px;
         border-radius: 10px;
     }
